app/app-Backend/requirements.txt
fastapi==0.100.1
uvicorn[standard]==0.23.2
motor==3.3.2
pymongo==4.6.0
python-dotenv==1.0.0
pydantic==1.10.12
email-validator==2.0.0
PyJWT==2.8.0
python-multipart==0.0.6
cloudinary==1.36.0
redis==5.0.1
werkzeug==3.0.1
stripe==5.5.0